Move edit modal into calendar.tsx
